+++ src/database/queries/bulkUploadTasks/insertOne.sql
@@ -1,5 +1,6 @@
 INSERT INTO bulk_upload_tasks (
 	source_id,
+	application_form_id,
 	funder_short_code,
 	proposals_data_file_id,
 	attachments_archive_file_id,
@@ -8,6 +9,7 @@ INSERT INTO bulk_upload_tasks (
 )
 VALUES (
 	:sourceId,
+	:applicationFormId,
 	:funderShortCode,
 	:proposalsDataFileId,
 	:attachmentsArchiveFileId,